+++ conf/script/build_system/compiler/host/get_info/version/__init__.py
@@ -1,1 +1,1 @@
-from build_system.compiler.host.get_info.version.fetch_compiler_version import assure_compiler_path_integrity, interpret_fetched_compiler_version, fetch_compiler_version
+from build_system.compiler.host.get_info.version.compiler import assure_compiler_path_integrity, interpret_fetched_compiler_version, fetch_compiler_version
