+++ conf/script/build_system/compiler/host/get_info/version/compiler.py
